Further refactored

+++ NCryptor.GUI/Helpers/FileServicesImpl.cs
@@ -14,20 +14,6 @@ namespace NCryptor.GUI.Helpers
             _options = fileOptions;
         }
 
-        public string ChangeOutputFileNameIfExists(string filePath)
-        {
-            if (!File.Exists(filePath))
-            {
-                return filePath;
-            }
-
-            var ext = Path.GetExtension(filePath);
-            var file = Path.GetFileNameWithoutExtension(filePath);
-
-            file += $" ({DateTime.Now})";
-            return Path.ChangeExtension(file, ext);
-        }
-
         public bool CheckFileExistance(string filePath)
         {
             return File.Exists(filePath);
@@ -35,25 +21,42 @@ namespace NCryptor.GUI.Helpers
 
         public string CreateDecryptedFilePath(string originalPath, string outputDirectory)
         {
-            var name = Path.GetFileNameWithoutExtension(Path.GetFileName(originalPath));
+            var name = Path.GetFileNameWithoutExtension(originalPath);
             var fullPath = Path.Combine(outputDirectory, name);
-            fullPath = ChangeOutputFileNameIfExists(fullPath);
 
-            return fullPath;
+            if (!CheckFileExistance(fullPath)) return fullPath;
+
+            return RandomlyChangeFileName(fullPath);
         }
 
         public string CreateEncryptedFilePath(string originalPath, string outputDirectory)
         {
-            var name = Path.GetFileName(originalPath);
-            name += _options.Extension;
-            name = ChangeOutputFileNameIfExists(name);
+            var ext = Path.GetExtension(originalPath);
+            var nameWithoutExt = Path.GetFileNameWithoutExtension(originalPath);
+            var nameWithExt = Path.GetFileName(originalPath);
+            var tempName = Path.Combine(outputDirectory, nameWithExt);
 
-            return Path.Combine(outputDirectory, name);
+            if(!CheckFileExistance($"{tempName}{_options.Extension}")) return $"{tempName}{_options.Extension}";
+
+            tempName = RandomlyChangeFileName(tempName);
+            return $"{tempName}{_options.Extension}";
         }
 
-        public void DeleteFile(string filePath)
+        public void DeleteFileIfExists(string filePath)
         {
-            File.Delete(filePath);
+            if (CheckFileExistance(filePath)) File.Delete(filePath);
+        }
+
+        private string RandomlyChangeFileName(string filePath)
+        {
+            var ext = Path.GetExtension(filePath);
+            var nameWithoutExt = Path.GetFileNameWithoutExtension(filePath);
+            var directory = Path.GetDirectoryName(filePath);
+
+            var fullPath = $"{nameWithoutExt} ({DateTime.Now: yyyy-MM-dd--HH-mm-ss})";
+            if (!string.IsNullOrEmpty(ext)) fullPath = $"{fullPath}{ext}";
+
+            return Path.Combine(directory, fullPath);
         }
     }
 }
